Move the implementation of the paths back into a single file, and move databrickspath to _posixpath.
Linting fixes are also reverted.
