--- src/databricks/labs/blueprint/paths/__init__.py
@@ -1,4 +1,0 @@
-from databricks.labs.blueprint.paths.dbfs import DBFSPath
-from databricks.labs.blueprint.paths.workspace import WorkspacePath
-
-__all__ = ("DBFSPath", "WorkspacePath")
